requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiolimiter==1.1.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dacite==1.8.1
frozenlist==1.3.3
geojson-pydantic==0.6.1
idna==3.4
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
plotly==5.14.1
protobuf==4.21.12
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
shapely==2.0.1
six==1.16.0
tenacity==8.2.2
traveltimepy==3.2.2
typing_extensions==4.5.0
tzdata==2023.3
yarl==1.9.2
